+++ CHANGELOG.md
@@ -1,9 +1,20 @@
 # Changelog
 
-## v0.14 - 2019-08-21
+## v0.15.0 - 2019-08-22
+
+* fixed error in `Genie.newapp()` with `dbsupport = true`
+* internal API cleanup and optimisations
+* fixed issue with `newresource` SearchLight integration
+* SearchLight initializer code is now uncommented
+* dependencies update
+* `Router.tolink` and its alias `Router.linkto` throw exceptions if the route is not defined **breaking**
+* `Router.tolink!!` and its alias `Router.linkto!!` have been removed **breaking**
+* new method `Requests.read(HttpFile, Type{String})` which returns the content of an uploaded file as a string.
+
+## v0.14.0 - 2019-08-21
 
 * consolidation of the Generator API
-* Genie depencencies update
+* Genie dependencies update
 * support for Julia v1.2
 * removal of the `REPL` module
 * CORS handling improvement (thanks @milesfrain)